Updated with Alex M.'s changes.
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@64 d25b4930-3292-e311-b074-0010c6de5da8

+++ ported-apps/Util/TimeMon-4.0/Percentages.h
@@ -10,7 +10,7 @@
 #import <AppKit/AppKit.h>
 #import "loadave.h"
 
-typedef long CPUTime[ CPUSTATES];
+typedef unsigned long long CPUTime[CPUSTATES];
 
 @interface Percentages : NSObject
 {
